# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "devops运维管理系统API接口文档",
        "title": "devops运维管理系统",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/clusters/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据提供的配置信息创建一个新的Kubernetes集群",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8s集群列表"
                ],
                "summary": "创建Kubernetes集群",
                "parameters": [
                    {
                        "description": "集群配置信息",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddKubeClusterDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功返回结果",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/clusters/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据提供的更新数据修改已有Kubernetes集群的配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8s集群列表"
                ],
                "summary": "更新Kubernetes集群信息",
                "parameters": [
                    {
                        "description": "更新后的集群配置信息",
                        "name": "cluster",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateKubeClusterDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功返回结果",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/v1/config/ecsauthadd": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Config配置中心"
                ],
                "summary": "创建凭据",
                "parameters": [
                    {
                        "description": "凭据信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateEcsPasswordAuthDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/config/ecsauthdelete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Config配置中心"
                ],
                "summary": "删除凭据",
                "parameters": [
                    {
                        "description": "凭据ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EcsAuthIdDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/config/ecsauthinfo": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Config配置中心"
                ],
                "summary": "根据名称获取凭据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "凭据名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EcsAuthVo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/config/ecsauthlist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Config配置中心"
                ],
                "summary": "获取所有凭据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.EcsAuthVo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/config/ecsauthupdate": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Config配置中心"
                ],
                "summary": "更新凭据",
                "parameters": [
                    {
                        "description": "凭据信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEcsAuthDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AddKubeClusterDto": {
            "type": "object",
            "required": [
                "kubeconfig",
                "name",
                "prometheus_url"
            ],
            "properties": {
                "isDefault": {
                    "description": "是否默认展示",
                    "type": "boolean"
                },
                "kubeconfig": {
                    "type": "string"
                },
                "name": {
                    "description": "集群名",
                    "type": "string"
                },
                "prometheus_url": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateEcsPasswordAuthDto": {
            "type": "object",
            "required": [
                "name",
                "password",
                "type",
                "username"
            ],
            "properties": {
                "name": {
                    "description": "凭证名称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "publicKey": {
                    "description": "公钥",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "type": {
                    "description": "认证类型:1-\u003e密码",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "model.EcsAuthIdDto": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID",
                    "type": "integer"
                }
            }
        },
        "model.EcsAuthVo": {
            "type": "object",
            "properties": {
                "createTime": {
                    "$ref": "#/definitions/util.HTime"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateEcsAuthDto": {
            "type": "object",
            "required": [
                "name",
                "password",
                "type",
                "username"
            ],
            "properties": {
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "凭证名称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "publicKey": {
                    "description": "公钥",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "type": {
                    "description": "认证类型:1-\u003e密码",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.UpdateKubeClusterDto": {
            "type": "object",
            "required": [
                "id",
                "kubeconfig"
            ],
            "properties": {
                "id": {
                    "description": "集群名",
                    "type": "integer"
                },
                "isDefault": {
                    "description": "是否默认展示",
                    "type": "boolean"
                },
                "kubeconfig": {
                    "type": "string"
                },
                "prometheus_url": {
                    "type": "string"
                }
            }
        },
...
        "result.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回的数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "util.HTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}